+++ docs/py/apps/marimo.py
@@ -33,7 +33,7 @@ df = pd.DataFrame({"x": [2, 1, 3]})
 ITable(df)
 
 # %% [markdown]
-# A Sample Marimo application is available at [`apps/marimo/widget.py`](https://github.com/mwouts/itables/tree/main/apps/marimo/widget.py).
+# A sample Marimo application is available at [`apps/marimo/widget.py`](https://github.com/mwouts/itables/tree/main/apps/marimo/widget.py).
 #
 # ⚠️ Some datatables buttons don't work when using the ITable widget within Marimo, see [issue #387](https://github.com/mwouts/itables/issues/387). The next paragraph provides a workaround.
 #
@@ -50,4 +50,4 @@ ITable(df)
 # mo.iframe(html)
 # ```
 #
-# A Sample Marimo application that uses `to_html_datatable` is available at [`apps/marimo/html.py`](https://github.com/mwouts/itables/tree/main/apps/marimo/html.py).
+# A sample Marimo application that uses `to_html_datatable` is available at [`apps/marimo/html.py`](https://github.com/mwouts/itables/tree/main/apps/marimo/html.py).