+++ tests/test_sample_pandas_dfs.py
@@ -1,24 +1,39 @@
 import json
 
-import pandas as pd
 import pytest
-from packaging import version
 
 from itables import show, to_html_datatable
 from itables.datatables_format import _format_pandas_series, generate_encoder
-from itables.javascript import pd_style
-from itables.sample_dfs import (
-    COLUMN_TYPES,
-    PANDAS_VERSION_MAJOR,
-    generate_random_df,
-    generate_random_series,
-    get_countries,
-    get_dict_of_test_dfs,
-    get_dict_of_test_series,
-    get_indicators,
-    get_pandas_styler,
-    get_population,
-)
+
+try:
+    import pandas as pd
+except ImportError:
+    pd = None
+    pytest.skip("Pandas is not available", allow_module_level=True)
+else:
+    from itables.sample_pandas_dfs import (
+        COLUMN_TYPES,
+        generate_random_df,
+        generate_random_series,
+        get_countries,
+        get_dict_of_test_dfs,
+        get_dict_of_test_series,
+        get_indicators,
+        get_pandas_styler,
+        get_population,
+    )
+
+
+try:
+    import pandas.io.formats.style as pd_style
+except ImportError:
+    pd_style = None
+
+try:
+    import IPython  # noqa F401
+except ImportError:
+    show = to_html_datatable
+
 
 # Make sure that displaying a dataframe does not trigger a warning  #107
 pytestmark = [
@@ -26,11 +41,6 @@ pytestmark = [
     # Seen on the CI on Py38 and Py39
     pytest.mark.filterwarnings("ignore::ResourceWarning"),
 ]
-
-if PANDAS_VERSION_MAJOR < 2:
-    # DeprecationWarning: `cumproduct` is deprecated as of NumPy 1.25.0,
-    # and will be removed in NumPy 2.0. Please use `cumprod` instead.
-    pytestmark.append(pytest.mark.filterwarnings("ignore::DeprecationWarning"))
 
 
 def test_get_countries(connected, use_to_html):
@@ -55,14 +65,11 @@ def test_get_indicators(connected, use_to_html):
 
 
 @pytest.mark.skipif(
-    version.parse(pd.__version__) < version.parse("1.0"),
-    reason="TypeError: Cannot interpret '<attribute 'dtype' of 'numpy.generic' objects>' as a data type",
-)
-@pytest.mark.skipif(
     pd_style is None,
     reason="Missing optional dependency 'Jinja2'. DataFrame.style requires jinja2.",
 )
 def test_get_pandas_styler(connected):
+    pytest.importorskip("matplotlib")
     styler = get_pandas_styler()
     show(styler, connected=connected, allow_html=True)
 
@@ -71,18 +78,18 @@ def kwargs_remove_none(**kwargs):
     return {key: value for key, value in kwargs.items() if value is not None}
 
 
-def test_show_test_dfs(df, connected, use_to_html, lengthMenu, monkeypatch):
+def test_show_test_dfs(pd_df, connected, use_to_html, lengthMenu, monkeypatch):
     show(
-        df,
+        pd_df,
         connected=connected,
         use_to_html=use_to_html,
         **kwargs_remove_none(lengthMenu=lengthMenu),
     )
 
 
-def test_to_html_datatable(df, connected, use_to_html, lengthMenu, monkeypatch):
+def test_to_html_datatable(pd_df, connected, use_to_html, lengthMenu, monkeypatch):
     to_html_datatable(
-        df,
+        pd_df,
         connected=connected,
         use_to_html=use_to_html,
         **kwargs_remove_none(lengthMenu=lengthMenu),
@@ -107,6 +114,7 @@ def test_show_test_series(series_name, series, connected, use_to_html, monkeypat
 
 
 def test_show_df_with_duplicate_column_names(connected, use_to_html):
+    assert pd is not None
     df = pd.DataFrame({"a": [0], "b": [0.0], "c": ["str"]})
     df.columns = ["duplicated_name"] * 3
     show(df, connected=connected, use_to_html=use_to_html)
@@ -114,6 +122,7 @@ def test_show_df_with_duplicate_column_names(connected, use_to_html):
 
 @pytest.mark.parametrize("type", COLUMN_TYPES)
 def test_generate_random_series(type, rows=2000):
+    assert pd is not None
     x = generate_random_series(rows, type)
     assert isinstance(x, pd.Series)
     assert len(x.index) == rows